isolate mechanics db crew
