+++ techies/fixtures/mechanicsgencrew/refs/mechanics_db/resources.md
